added if clause and docs to ./git-remote-to-ssh.sh

+++ git-remote-to-ssh.sh
@@ -1,9 +1,26 @@
 #!/bin/bash
-# Usage: {remote name} {username} {repo name}
-git remote -v
-git remote rm $1
-git remote add $1 git@github.com:$2/$3
 
-echo ""
-echo "Changed to:"
-git remote -v
+if [ -z "$1" ] && [ -z "$2" ]; then
+    echo "Usage: <remote name> <username> <repo name>"
+    echo "example: ./git-remote-to-ssh.sh origin kushalj scripts"
+
+    echo ""
+    echo "Will do:"
+    echo "
+    origin  git@github.com:kushalj/scripts.git (fetch)
+    origin  git@github.com:kushalj/scripts.git (push)
+
+    Changed to:
+    origin  git@github.com:kushalj/scripts.git (fetch)
+    origin  git@github.com:kushalj/scripts.git (push)
+    "
+else
+    git remote -v
+    git remote rm $1
+    git remote add $1 git@github.com:$2/$3.git
+
+    echo ""
+    echo "Changed to:"
+    git remote -v
+fi
+
